fix: extract component
