+++ frontend/src/component/feature/FeatureView/FeatureOverview/ReleasePlan/ReleasePlanMilestone/MilestoneAutomationSection.tsx
@@ -1,0 +1,79 @@
+import Add from '@mui/icons-material/Add';
+import { Button, styled } from '@mui/material';
+import type { MilestoneStatus } from './ReleasePlanMilestoneStatus.tsx';
+
+const StyledAutomationContainer = styled('div', {
+    shouldForwardProp: (prop) => prop !== 'status',
+})<{ status?: MilestoneStatus }>(({ theme, status }) => ({
+    border: `1px solid ${status === 'active' ? theme.palette.success.border : theme.palette.divider}`,
+    borderTop: `1px solid ${theme.palette.divider}`,
+    borderRadius: `0 0 ${theme.shape.borderRadiusLarge}px ${theme.shape.borderRadiusLarge}px`,
+    padding: theme.spacing(1.5, 2),
+    backgroundColor: theme.palette.background.paper,
+    display: 'flex',
+    flexDirection: 'column',
+    alignItems: 'stretch',
+    gap: theme.spacing(1),
+    '& > *': {
+        alignSelf: 'flex-start',
+    },
+}));
+
+const StyledAddAutomationButton = styled(Button)(({ theme }) => ({
+    textTransform: 'none',
+    fontWeight: theme.typography.fontWeightBold,
+    fontSize: theme.typography.body2.fontSize,
+    padding: 0,
+    minWidth: 'auto',
+    gap: theme.spacing(1),
+    '&:hover': {
+        backgroundColor: 'transparent',
+    },
+    '& .MuiButton-startIcon': {
+        margin: 0,
+        width: 20,
+        height: 20,
+        border: `1px solid ${theme.palette.primary.main}`,
+        backgroundColor: theme.palette.background.elevation2,
+        borderRadius: '50%',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+        '& svg': {
+            fontSize: 14,
+            color: theme.palette.primary.main,
+        },
+    },
+}));
+
+interface IMilestoneAutomationSectionProps {
+    showAutomation?: boolean;
+    status?: MilestoneStatus;
+    onAddAutomation?: () => void;
+    automationForm?: React.ReactNode;
+}
+
+export const MilestoneAutomationSection = ({
+    showAutomation,
+    status,
+    onAddAutomation,
+    automationForm,
+}: IMilestoneAutomationSectionProps) => {
+    if (!showAutomation) return null;
+
+    return (
+        <StyledAutomationContainer status={status}>
+            {automationForm ? (
+                automationForm
+            ) : (
+                <StyledAddAutomationButton
+                    onClick={onAddAutomation}
+                    color='primary'
+                    startIcon={<Add />}
+                >
+                    Add automation
+                </StyledAddAutomationButton>
+            )}
+        </StyledAutomationContainer>
+    );
+};
